Combine notification classes
Also added missing priv check for rejecting applications.
